pyproject.toml
[project]
name = "cbdb"
version = "1.0.0"
description = "Browser-based Calibre e-book library viewer using WASM SQLite"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pillow>=12.1.0",
    "pytest>=9.0.2",
    "pytest-playwright>=0.7.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
